backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-generativeai
qdrant-client
sentence-transformers
scrapy
python-dotenv
